# Optimizing Xdebug Performance When Debugging is Inactive