When using Xdebug, enabling **debug mode** significantly impacts execution speed, often **doubling the run time** of code compared to when Xdebug is not loaded. This performance penalty occurs **even if the debugger is not actively used** —i.e., when no client connects or no breakpoints are set. As a result, developers often find themselves frequently enabling and disabling Xdebug to avoid noticeable slowdowns.

## Cause of the Slowdown
This slowdown happens because Xdebug injects additional processing at two critical points:

1. **On each function call** – It records stack traces, including function calls and variable values.
2. **On each statement execution** – It checks whether execution should pause at a breakpoint.

These operations are essential when debugging but become **unnecessary overhead** when the debugger is inactive.

## What This PR Changes
This PR introduces optimizations that **dynamically disable unnecessary debugging operations** when:

- Xdebug is running **only in debug mode** (i.e., without code coverage, profiling, tracing, etc...).
- The debugger is **not activated**, meaning:
  - The debugger is not started,
  - Or no client is connected,
  - Or no breakpoints are set.

By skipping function call logging and statement checks in these cases, we can **greatly reduce execution overhead** while maintaining full debugging capabilities when needed.

## Handling Just-In-Time (JIT) Debugging
One challenge is **JIT debugging**, where the debugger is automatically started when:

- An **exception or error** occurs.
- `xdebug_break()` is called.
- `xdebug_connect_to_client()` is used.

In these cases, some debugging functionality must remain available to ensure Xdebug can correctly **attach to the process and capture execution state**. As a result, the optimizations in this PR do **not** entirely eliminate Xdebug’s overhead but **significantly reduce it** compared to the current implementation.

## Challenges
A key challenge is that **JIT debugging sessions will not have a pre-recorded stack trace**. However, this is an issue thar can be overcome because:

- The stack trace can be reconstructed when a debugger client connects.
- While reconstructing the stack trace incurs some performance cost, it happens **only once during reconnection**, rather than continuously during execution.

This is a reasonable compromise to ensure that Xdebug remains performant when debugging is not actively in use.

## Performance Improvements
To illustrate the benefits of this approach, we measured execution times for two common real-world use cases: Running **RectorPHP** and **PHPStan** on a codebase.

| **Operation**            | **Without Xdebug (s)** | **With Current Xdebug (s)** | **Slowdown (%)** | **With Optimized Xdebug (s)** | **Slowdown (%)** |
|--------------------------|----------------------|----------------------|----------------|----------------------|----------------|
| Running **RectorPHP**    | 42                   | 92                   | **+119%**       | **52**                | **+23%**       |
| Running **PHPStan**      | 51                   | 108                  | **+112%**       | **62**                | **+22%**       |

These tests were conducted with **Xdebug enabled, a debugger client connected, and no breakpoints set**. As the results show, the optimized implementation **runs 3–4× faster** than the current version when debugging is not actively used.

## Goal of This Change
The ultimate goal of this improvement is to **make Xdebug’s performance impact minimal when debugging is inactive**. This means developers no longer need to constantly toggle Xdebug on and off. Instead, they can leave it enabled at all times with:

```ini
xdebug.mode=debug
xdebug.start_with_request=yes
```

This allows seamless debugging when needed, without causing unnecessary slowdowns during regular execution.

## Additional improvements

This PR also includes a few optimizations that enhance the execution speed of the debugger, regardless of whether it is actively used or not. These improvements further reduce the overall overhead of Xdebug, making debugging more efficient while maintaining full functionality.

## Support This Work

If you or your company use Xdebug, please consider supporting my open-source work by [sponsoring me](https://github.com/sponsors/carlos-granados) on GitHub. Your support helps fund further improvements, and I have several ideas for enhancing Xdebug even further.